testing/requirements.txt
allure-pytest==2.8.32
allure-python-commons==2.8.32
attrs==21.2.0
execnet==1.9.0
importlib-metadata==4.8.1
iniconfig==1.1.1
Faker==8.0.0
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==2.4.0
requests==2.25.1
selenium==4.0.0a5
toml==0.10.2
typing-extensions==3.10.0.2
zipp==3.6.0
PyMySQL==1.0.2
SQLAlchemy==1.4.13